fix(bench): satisfy clippy without allows
cast_precision_loss on the per-name throughput: convert the closure
count through u32 (exact in f64) instead of a raw usize cast. The
exit-code constants become u8 — ExitCode::from is their only numeric
consumer, which also drops the hosts' `as u8` casts.

+++ crates/bench/src/lib.rs
@@ -29,9 +29,9 @@ use std::path::Path;
 use serde_json::{Map, Value};
 
 /// Bad invocation (unknown name, missing input, conflicting flags).
-pub const EXIT_USAGE: i32 = 2;
+pub const EXIT_USAGE: u8 = 2;
 /// The kernel rejected at least one constant; its row is on disk.
-pub const EXIT_REJECTED: i32 = 3;
+pub const EXIT_REJECTED: u8 = 3;
 
 /// Row outcome as written by a tool (`Oom` is only ever set by the
 /// orchestrator, after the fact — a tool never observes its own OOM kill).
